refactor to class
